bug when adding / checking user
When checking if user data exists, if the data was found it pretended it wasn't, and if it wasn't then the bot wouldn't add them.
Also removed some of the extra user data entries created by the bug.
Added users.json and token.json to .gitignore so they will no longer update