updated *col command ( fixed bug where *col wouldn't send)
Sometimes the size of user collections was too large, and the output message would exceed the 500 character message limit set by twitch and therefor would not send, if messages are above 500 characters it will now send in 2 seperate messages.
If the scale increases more, then this will have to be updated again to allow for 3 messages.